simplified upgrade command split upgrade API
