+++ cmd/upgrade.go
@@ -8,7 +8,6 @@ import (
 	"path"
 
 	"github.com/Azure/acs-engine/pkg/api"
-	"github.com/Azure/acs-engine/pkg/api/vlabs"
 	"github.com/Azure/acs-engine/pkg/armhelpers"
 	"github.com/Azure/acs-engine/pkg/i18n"
 	"github.com/Azure/acs-engine/pkg/operations/kubernetesupgrade"
@@ -30,7 +29,7 @@ type upgradeCmd struct {
 	// user input
 	resourceGroupName   string
 	deploymentDirectory string
-	upgradeModelFile    string
+	upgradeVersion      string
 	containerService    *api.ContainerService
 	apiVersion          string
 
@@ -57,7 +56,7 @@ func newUpgradeCmd() *cobra.Command {
 	f := upgradeCmd.Flags()
 	f.StringVar(&uc.resourceGroupName, "resource-group", "", "the resource group where the cluster is deployed")
 	f.StringVar(&uc.deploymentDirectory, "deployment-dir", "", "the location of the output from `generate`")
-	f.StringVar(&uc.upgradeModelFile, "upgrademodel-file", "", "file path to upgrade API model")
+	f.StringVar(&uc.upgradeVersion, "upgrade-version", "", "desired kubernetes version")
 	addAuthFlags(&uc.authArgs, f)
 
 	return upgradeCmd
@@ -79,9 +78,9 @@ func (uc *upgradeCmd) validate(cmd *cobra.Command, args []string) {
 	}
 
 	// TODO(colemick): add in the cmd annotation to help enable autocompletion
-	if uc.upgradeModelFile == "" {
+	if uc.upgradeVersion == "" {
 		cmd.Usage()
-		log.Fatal("--upgrademodel-file must be specified")
+		log.Fatal("--upgrade-version must be specified")
 	}
 
 	if uc.client, err = uc.authArgs.getClient(); err != nil {
@@ -109,17 +108,28 @@ func (uc *upgradeCmd) validate(cmd *cobra.Command, args []string) {
 	if err != nil {
 		log.Fatalf("error parsing the api model: %s", err.Error())
 	}
-	if _, err = os.Stat(uc.upgradeModelFile); os.IsNotExist(err) {
-		log.Fatalf("specified upgrade model file does not exist (%s)", uc.upgradeModelFile)
-	}
 
-	// validate upgrade and set the Goal State
-	contents, err := ioutil.ReadFile(uc.upgradeModelFile)
+	// get available upgrades for container service
+	orchestratorInfo, err := api.GetOrchestratorVersionProfile(uc.containerService.Properties.OrchestratorProfile)
 	if err != nil {
-		log.Fatalf("error reading file %s: %s", uc.upgradeModelFile, err.Error())
+		log.Fatalf("error getting list of available upgrades: %s", err.Error())
 	}
-	if err = apiloader.UpdateContainerServiceForUpgrade(contents, vlabs.APIVersion, uc.containerService, true); err != nil {
-		log.Fatalf("error loading ContainerService: %s", err.Error())
+	// add the current version if upgrade has failed
+	orchestratorInfo.Upgrades = append(orchestratorInfo.Upgrades, &api.OrchestratorProfile{
+		OrchestratorType:    uc.containerService.Properties.OrchestratorProfile.OrchestratorType,
+		OrchestratorVersion: uc.containerService.Properties.OrchestratorProfile.OrchestratorVersion})
+
+	// validate desired upgrade version and set goal state
+	found := false
+	for _, up := range orchestratorInfo.Upgrades {
+		if up.OrchestratorVersion == uc.upgradeVersion {
+			uc.containerService.Properties.OrchestratorProfile.OrchestratorVersion = uc.upgradeVersion
+			found = true
+			break
+		}
+	}
+	if !found {
+		log.Fatalf("version %s is not supported", uc.upgradeVersion)
 	}
 
 	uc.client, err = uc.authArgs.getClient()
@@ -133,7 +143,7 @@ func (uc *upgradeCmd) validate(cmd *cobra.Command, args []string) {
 	// user could have specified a name suffix instead of using the default
 	// value generated by ACS Engine
 	templatePath := path.Join(uc.deploymentDirectory, "azuredeploy.json")
-	contents, _ = ioutil.ReadFile(templatePath)
+	contents, _ := ioutil.ReadFile(templatePath)
 
 	var template interface{}
 	json.Unmarshal(contents, &template)
