+++ examples/k8s-upgrade/README.md
@@ -5,8 +5,8 @@
 This document describes how to upgrade kubernetes version on a running cluster.
 
 Supported scenarios:
-- upgrade from v1.5.x to v1.6.11
-- upgrade from v1.6.x to v1.7.7
+- upgrade from v1.5.x to the latest supported version in v1.6 stream
+- upgrade from v1.6.x to the latest supported version in v1.7 stream
 
 The cluster definition file examples demonstrate initial cluster configurations:
 - **kubernetes1.5.json** - Kubernetes cluster v1.5 with Linux agent pool
