+++ examples/k8s-upgrade/k8s-upgrade.sh
@@ -11,20 +11,12 @@ fi
 [[ ! -z "${EXPECTED_ORCHESTRATOR_VERSION:-}" ]] || (echo "Must specify EXPECTED_ORCHESTRATOR_VERSION" && exit 1)
 
 OUTPUT="_output/${INSTANCE_NAME}"
-K8S_UPGRADE_CONF="$OUTPUT/k8sUpgrade.json"
-
-cat > $K8S_UPGRADE_CONF <<END
-{
-  "orchestratorType": "Kubernetes",
-  "orchestratorVersion": "${EXPECTED_ORCHESTRATOR_VERSION}"
-}
-END
 
 ./bin/acs-engine upgrade \
   --subscription-id ${SUBSCRIPTION_ID} \
   --deployment-dir ${OUTPUT} \
   --resource-group ${RESOURCE_GROUP} \
-  --upgrademodel-file $K8S_UPGRADE_CONF \
+  --upgrade-version ${EXPECTED_ORCHESTRATOR_VERSION} \
   --auth-method client_secret \
   --client-id ${SERVICE_PRINCIPAL_CLIENT_ID} \
   --client-secret ${SERVICE_PRINCIPAL_CLIENT_SECRET}
