+++ pkg/api/agentPoolOnlyApi/v20170831/types.go
@@ -94,6 +94,26 @@ const (
 	Migrating ProvisioningState = "Migrating"
 )
 
+// PoolUpgradeProfile contains pool properties:
+//  - kubernetes version
+//  - pool name (for agent pool)
+//  - OS type of the VMs in the pool
+//  - list of applicable upgrades
+type PoolUpgradeProfile struct {
+	KubernetesVersion string    `json:"kubernetesVersion"`
+	Name              string    `json:"name,omitempty"`
+	OSType            string    `json:"osType,omitempty"`
+	Upgrades          []*string `json:"upgrades,omitempty"`
+}
+
+// UpgradeProfile contains cluster properties:
+//  - orchestrator type and version for the cluster
+//  - list of pool profiles, constituting the cluster
+type UpgradeProfile struct {
+	ControlPlaneProfile *PoolUpgradeProfile   `json:"controlPlaneProfile"`
+	AgentPoolProfiles   []*PoolUpgradeProfile `json:"agentPoolProfiles"`
+}
+
 // AgentPoolProfile represents configuration of VMs running agent
 // daemons that register with the master and offer resources to
 // host applications in containers.
