+++ pkg/api/apiloader.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/Azure/acs-engine/pkg/api/agentPoolOnlyApi/v20170831"
 	apvlabs "github.com/Azure/acs-engine/pkg/api/agentPoolOnlyApi/vlabs"
-	"github.com/Azure/acs-engine/pkg/api/upgrade/v20170930"
 	"github.com/Azure/acs-engine/pkg/api/v20160330"
 	"github.com/Azure/acs-engine/pkg/api/v20160930"
 	"github.com/Azure/acs-engine/pkg/api/v20170131"
@@ -181,60 +180,6 @@ func (a *Apiloader) LoadContainerServiceForAgentPoolOnlyCluster(contents []byte,
 	default:
 		return nil, a.Translator.Errorf("unrecognized APIVersion in LoadContainerServiceForAgentPoolOnlyCluster '%s'", version)
 	}
-}
-
-// UpdateContainerServiceForUpgrade pre-validates upgrade operation and updates container service
-func (a *Apiloader) UpdateContainerServiceForUpgrade(
-	contents []byte,
-	version string,
-	cs *ContainerService,
-	allowCurrentVersionUpgrade bool) error {
-	unverOrch := &OrchestratorProfile{}
-
-	switch version {
-	case v20170930.APIVersion:
-		up := &v20170930.OrchestratorProfile{}
-		if e := json.Unmarshal(contents, up); e != nil {
-			return a.Translator.Errorf(e.Error())
-		}
-		if e := up.ValidateForUpgrade(); e != nil {
-			return a.Translator.Errorf(e.Error())
-		}
-		convertV20170930OrchestratorProfile(up, unverOrch)
-
-	case vlabs.APIVersion:
-		up := &vlabs.OrchestratorProfile{}
-		if e := json.Unmarshal(contents, up); e != nil {
-			return a.Translator.Errorf(e.Error())
-		}
-		if e := up.ValidateForUpgrade(); e != nil {
-			return a.Translator.Errorf(e.Error())
-		}
-		convertVLabsOrchestratorProfile(up, unverOrch)
-
-	default:
-		return a.Translator.Errorf("unrecognized APIVersion in UpdateContainerServiceForUpgrade '%s'", version)
-	}
-
-	// get available upgrades for container service
-	orchestratorInfo, e := GetOrchestratorVersionProfile(cs.Properties.OrchestratorProfile)
-	if e != nil {
-		return e
-	}
-	// add current version if upgrade has failed
-	if allowCurrentVersionUpgrade {
-		orchestratorInfo.Upgrades = append(orchestratorInfo.Upgrades, &OrchestratorProfile{
-			OrchestratorVersion: cs.Properties.OrchestratorProfile.OrchestratorVersion})
-	}
-	// validate desired upgrade version and set goal state
-	for _, up := range orchestratorInfo.Upgrades {
-		if up.OrchestratorVersion == unverOrch.OrchestratorVersion {
-			cs.Properties.OrchestratorProfile.OrchestratorVersion = up.OrchestratorVersion
-			return nil
-		}
-	}
-	return a.Translator.Errorf("Kubernetes %s cannot be upgraded to %s",
-		cs.Properties.OrchestratorProfile.OrchestratorVersion, unverOrch.OrchestratorVersion)
 }
 
 // SerializeContainerService takes an unversioned container service and returns the bytes
