+++ pkg/api/convertertoapi.go
@@ -4,7 +4,6 @@ import (
 	"strings"
 
 	"github.com/Azure/acs-engine/pkg/api/common"
-	"github.com/Azure/acs-engine/pkg/api/upgrade/v20170930"
 	"github.com/Azure/acs-engine/pkg/api/v20160330"
 	"github.com/Azure/acs-engine/pkg/api/v20160930"
 	"github.com/Azure/acs-engine/pkg/api/v20170131"
@@ -532,20 +531,6 @@ func convertV20170131OrchestratorProfile(v20170131 *v20170131.OrchestratorProfil
 	} else if api.OrchestratorType == DCOS {
 		api.OrchestratorVersion = DCOSVersion1Dot9Dot0
 	}
-}
-
-func convertV20170930OrchestratorProfile(v *v20170930.OrchestratorProfile, api *OrchestratorProfile) {
-	switch v.OrchestratorType {
-	case v20170930.Kubernetes:
-		api.OrchestratorType = Kubernetes
-	case v20170930.DCOS:
-		api.OrchestratorType = DCOS
-	case v20170930.Swarm:
-		api.OrchestratorType = Swarm
-	case v20170930.DockerCE:
-		api.OrchestratorType = SwarmMode
-	}
-	api.OrchestratorVersion = v.OrchestratorVersion
 }
 
 func convertV20170701OrchestratorProfile(v20170701cs *v20170701.OrchestratorProfile, api *OrchestratorProfile) {
