+++ pkg/api/v20170701/types.go
@@ -184,6 +184,26 @@ type AgentPoolProfile struct {
 	subnet string
 }
 
+// PoolUpgradeProfile contains pool properties:
+//  - orchestrator type and version
+//  - pool name (for agent pool)
+//  - OS type of the VMs in the pool
+//  - list of applicable upgrades
+type PoolUpgradeProfile struct {
+	OrchestratorProfile
+	Name     string                 `json:"name,omitempty"`
+	OSType   string                 `json:"osType,omitempty"`
+	Upgrades []*OrchestratorProfile `json:"upgrades,omitempty"`
+}
+
+// UpgradeProfile contains cluster properties:
+//  - orchestrator type and version for the cluster
+//  - list of pool profiles, constituting the cluster
+type UpgradeProfile struct {
+	ControlPlaneProfile *PoolUpgradeProfile   `json:"controlPlaneProfile"`
+	AgentPoolProfiles   []*PoolUpgradeProfile `json:"agentPoolProfiles"`
+}
+
 // UnmarshalJSON unmarshal json using the default behavior
 // And do fields manipulation, such as populating default value
 func (a *AgentPoolProfile) UnmarshalJSON(b []byte) error {
