+++ pkg/api/v20170930/types.go
@@ -12,26 +12,6 @@ type OrchestratorProfile struct {
 	OrchestratorVersion string `json:"orchestratorVersion"`
 }
 
-// PoolUpgradeProfile contains pool properties:
-//  - orchestrator type and version
-//  - pool name (for agent pool)
-//  - OS type of the VMs in the pool
-//  - list of applicable upgrades
-type PoolUpgradeProfile struct {
-	OrchestratorProfile
-	Name     string                 `json:"name,omitempty"`
-	OSType   string                 `json:"osType,omitempty"`
-	Upgrades []*OrchestratorProfile `json:"upgrades,omitempty"`
-}
-
-// UpgradeProfile contains cluster properties:
-//  - orchestrator type and version for the cluster
-//  - list of pool profiles, constituting the cluster
-type UpgradeProfile struct {
-	ControlPlaneProfile *PoolUpgradeProfile   `json:"controlPlaneProfile"`
-	AgentPoolProfiles   []*PoolUpgradeProfile `json:"agentPoolProfiles"`
-}
-
 // OrchestratorVersionProfile contains information of a supported orchestrator version:
 //  - orchestrator type and version
 //  - whether this orchestrator version is deployed by default if orchestrator release is not specified
